+++ integreat_cms/cms/forms/pages/page_translation_form.py
@@ -81,6 +81,9 @@ class PageTranslationForm(CustomContentModelForm):
 
         # Create new version if content changed
         if not {"slug", "title", "content"}.isdisjoint(self.changed_data):
+            # Delete now outdated link objects
+            self.instance.links.all().delete()
+            # Save new version
             self.instance.version += 1
             self.instance.pk = None