+++ integreat_cms/cms/templatetags/text_filters.py
@@ -38,7 +38,7 @@ def linkcheck_status_filter(status_message):
         return _("Unknown")
     if status_message.startswith("Other Error:"):
         return _("Error")
-    return status_message
+    return _(status_message)
 
 
 @register.filter(name="to_json")