+++ integreat_cms/cms/utils/__init__.py
@@ -1,3 +1,3 @@
 """
-This package contains utilities which faciliate specific tasks and can be re-used and imported from different locations.
+This package contains utilities which facilitate specific tasks and can be re-used and imported from different locations.
 """